Image Classification/requirements.txt
tensorflow==2.6.0
keras==2.6.0
numpy==1.19.5
Pillow==8.3.2
tk==0.1.0
scipy==1.7.2